Apply code review security and quality improvements
- Fix verifyKeyCompatibility to use constant-time comparison for cryptographic operations
- Add mpz_invert return value checks in decrypt, decryptCard, and cooperativeReveal
- Add bounds validation in Card::fromInt and decryptCard for card values
- Change permutation type from int to size_t for 64-bit safety
- Add shrink_to_fit for S_matrix clearing to free memory
- Add validation for decrypted values in decryptCard function

+++ apps/main.cpp
@@ -73,7 +73,7 @@ int main() {
     std::cout << "\n=== Alice's Shuffle ===" << std::endl;
 
     size_t n = deck.size();
-    std::vector<int> perm1 = BayerGrothShuffle::generatePermutation(n, rng1);
+    std::vector<size_t> perm1 = BayerGrothShuffle::generatePermutation(n, rng1);
 
     std::vector<mpz_class> rand1(n);
     for (size_t i = 0; i < n; ++i) {
@@ -111,7 +111,7 @@ int main() {
     auto reencryptDuration = std::chrono::duration_cast<std::chrono::microseconds>(reencryptEnd - reencryptStart);
 
     std::cout << "Shuffling deck..." << std::endl;
-    std::vector<int> perm2 = BayerGrothShuffle::generatePermutation(n, rng2);
+    std::vector<size_t> perm2 = BayerGrothShuffle::generatePermutation(n, rng2);
 
     std::vector<mpz_class> rand2(n);
     for (size_t i = 0; i < n; ++i) {
